requirements.txt
Flask==2.3.2
openpyxl==3.1.2
gunicorn==20.1.0
psycopg2-binary
reportlab